bucket limiter

+++ pkg/rate_limitier/bucket/cmd/limitier/bucket.go
@@ -27,10 +27,10 @@ type bucket struct {
 // this operation happens on a request level
 // avoiding long running term routine
 func (b *bucket) refill(nowTime time.Time, size int, rate int) error {
-	elaspsed := int(nowTime.Sub(b.lastRefilled).Seconds())
-	secRate := rate / 60
+	elaspsed := nowTime.Sub(b.lastRefilled).Seconds()
+	secRate := float64(rate / 60)
 
-	refill := elaspsed * secRate
+	refill := int(elaspsed * secRate)
 
 	if b.currentUnits+refill >= size {
 		b.currentUnits = size
